input is working
