+++ src/components/searchAndSlider.js
@@ -1,95 +1,124 @@
-'use strict';
-import Siema from 'siema';
+// 'use strict';
+// import Siema from 'siema';
 
-const refs = {
-  input: document.querySelector('.d_search-box'),
-  starBtn: document.querySelector('.d_star-icon'),
-  cityContainer: document.querySelector('.d_siema'),
-  deleteBtn: document.querySelector('.d_close-btn'),
-};
+// const refs = {
+//   input: document.querySelector('.d_search-box'),
+//   starBtn: document.querySelector('.d_star-icon'),
+//   cityContainer: document.querySelector('.d_siema'),
+//   deleteBtn: document.querySelector('.d_close-btn'),
+// };
 
-const citiesFromStorage = JSON.parse(localStorage.getItem('City'));
-const savedCities = [];
+// const citiesFromStorage = JSON.parse(localStorage.getItem('City'));
+// const savedCities = [];
 
-if (citiesFromStorage === null) {
-  const allItems = savedCities
-    .map(
-      city => `<div id=${city} class="d_btn-container">
-        <span class="d_added-city">${city}</span>
-        <button class="d_close-btn" data-id=${city}></button>
-        </div>`,
-    )
-    .join('');
-  refs.cityContainer.innerHTML = allItems;
-} else {
-  const allItems = citiesFromStorage
-    .map(
-      city => `<div id=${city} class="d_btn-container"><span class="d_added-city">${city}</span><button class="d_close-btn"
-        data-id=${city}></button>
-      </div>`,
-    )
-    .join('');
-  refs.cityContainer.innerHTML = allItems;
-}
+// if (citiesFromStorage === null) {
+//   const allItems = savedCities
+//     .map(
+//       city => `<div id=${city} class="d_btn-container">
+//         <span class="d_added-city">${city}</span>
+//         <button class="d_close-btn" data-id=${city}></button>
+//         </div>`,
+//     )
+//     .join('');
+//   refs.cityContainer.innerHTML = allItems;
+// } else {
+//   const allItems = citiesFromStorage
+//     .map(
+//       city => `<div id=${city} class="d_btn-container"><span class="d_added-city">${city}</span><button class="d_close-btn"
+//         data-id=${city}></button>
+//       </div>`,
+//     )
+//     .join('');
+//   refs.cityContainer.innerHTML = allItems;
+// }
 
-refs.starBtn.addEventListener('click', insertCityToTheMem);
-document.addEventListener('click', deleteCityFromMem);
+// let citiesMemList = null;
+// refs.input.addEventListener('input', debounce(searchFromInput, 1000));
+// refs.starBtn.addEventListener('click', insertCityToTheMem);
+// document.addEventListener('click', deleteCityFromMem);
 
-function insertCityToTheMem() {
-  refs.starBtn.classList.add('d_full-star');
-  let { value } = refs.input;
+// function searchFromInput(e) {
+//   let city = e.target.value;
 
-  if (!value || savedCities.includes(value)) {
-    return;
-  }
-  refs.input.value = '';
+//   query
+//     .fetchWeatherForecastByCity(city)
+//     .then(data => {
+//       let keys = Object.values(data.sys.country).join('');
 
-  savedCities.push(value);
+//       if (data.cod === 200 && !citiesFromStorage(data.name)) {
+//         refs.starBtn.classList.remove('d_full-star');
+//         pixabayApi.searchQuerry = data.name;
 
-  const allItems = savedCities
-    .map(
-      city => `<div id=${city} class="d_btn-container">
-        <span class="d_added-city">${city}</span>
-        <button class="d_close-btn" data-id=${city}></button>
-        </div>`,
-    )
-    .join('');
+//         citiesMemList = {
+//           name: data.name,
+//           html: memoryBuilder(data),
+//         };
+//       } else {
+//         citiesMemList = null;
+//       }
+//     })
+//     .catch(err => {
+//       const myErr = error({
+//         delay: 2000,
+//         text: 'Sorry! We can`t find your reqest.',
+//       });
+//     });
+// }
+// function insertCityToTheMem() {
+//   refs.starBtn.classList.add('d_full-star');
+//   let { value } = refs.input;
 
-  refs.cityContainer.innerHTML = allItems;
+//   if (!value || savedCities.includes(value)) {
+//     return;
+//   }
+//   refs.input.value = '';
 
-  // LOCAL STORAGE
-  localStorage.setItem('City', JSON.stringify(savedCities));
+//   savedCities.push(value);
 
-  const mySiema = new Siema({
-    perPage: findedCities.length > 3 ? 3 : findedCities.length,
-    multipleDrag: true,
-    selector: '.siema',
-    easing: 'ease-out',
-    draggable: true,
-    duration: 200,
-    threshold: 20,
-    startIndex: 0,
-    loop: true,
-    rtl: false,
-    onInit: () => {},
-    onChange: () => {},
-  });
+//   const allItems = savedCities
+//     .map(
+//       city => `<div id=${city} class="d_btn-container">
+//         <span class="d_added-city">${city}</span>
+//         <button class="d_close-btn" data-id=${city}></button>
+//         </div>`,
+//     )
+//     .join('');
 
-  document
-    .querySelector('.d_right-btn')
-    .addEventListener('click', () => mySiema.next());
-}
+//   refs.cityContainer.innerHTML = allItems;
 
-function deleteCityFromMem(e) {
-  if (e.target && e.target.className === 'd_close-btn') {
-    let id = e.target.dataset.id;
-    document.getElementById(id).remove();
-    savedCities.splice(savedCities.indexOf(id), 1);
+//   // LOCAL STORAGE
+//   localStorage.setItem('City', JSON.stringify(savedCities));
 
-    const cutItem = citiesFromStorage.splice(citiesFromStorage.indexOf(id), 1);
-    localStorage.setItem('City', JSON.stringify(cutItem));
-    if ((citiesFromStorage.length = 1)) {
-      localStorage.clear();
-    }
-  }
-}
+//   const mySiema = new Siema({
+//     perPage: citiesFromStorage.length > 3 ? 3 : citiesFromStorage.length,
+//     multipleDrag: true,
+//     selector: '.siema',
+//     easing: 'ease-out',
+//     draggable: true,
+//     duration: 200,
+//     threshold: 20,
+//     startIndex: 0,
+//     loop: true,
+//     rtl: false,
+//     onInit: () => {},
+//     onChange: () => {},
+//   });
+
+//   document
+//     .querySelector('.d_right-btn')
+//     .addEventListener('click', () => mySiema.next());
+// }
+
+// function deleteCityFromMem(e) {
+//   if (e.target && e.target.className === 'd_close-btn') {
+//     let id = e.target.dataset.id;
+//     document.getElementById(id).remove();
+//     savedCities.splice(savedCities.indexOf(id), 1);
+
+//     const cutItem = citiesFromStorage.splice(citiesFromStorage.indexOf(id), 1);
+//     localStorage.setItem('City', JSON.stringify(cutItem));
+//     if ((citiesFromStorage.length = 1)) {
+//       localStorage.clear();
+//     }
+//   }
+// }
